requirements.txt
apache-airflow[postgres]
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres